requirements.txt
Flask>=2.0
python-dotenv>=0.19
google-api-python-client>=2.0
newspaper3k>=0.2.8
google-generativeai>=0.6.0

requests>=2.25
readability-lxml>=0.8
trafilatura>=1.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
